Revert "feat: add welcome page (#4)"
This reverts commit 837485238dd5c7496994b1869758d33955724b49.
