+++ client/package.json
@@ -10,16 +10,11 @@
   },
   "dependencies": {
     "@apollo/client": "^3.11.10",
-    "@gorhom/bottom-sheet": "^5",
     "expo": "~52.0.6",
     "expo-status-bar": "~2.0.0",
     "graphql": "^16.9.0",
     "react": "18.3.1",
     "react-native": "0.76.1",
-    "react-native-css-interop": "^0.1.22",
-    "react-native-gesture-handler": "~2.20.2",
-    "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "^4.14.0",
     "react-native-sketchbook": "^0.1.1"
   },
   "devDependencies": {
